services/user-service/requirements.txt
fastapi==0.85.0
uvicorn==0.17.6
sqlalchemy==1.4.39
psycopg2-binary==2.9.5
pydantic==1.10.2
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
alembic==1.8.1
pytest==7.1.2
